requirements.txt
# if using GPU and CUDA 11.8 then add --index-url https://download.pytorch.org/whl/cu118
# see (https://pytorch.org/get-started/locally/)
torch==2.1.2        # --index-url https://download.pytorch.org/whl/cu118
torchvision==0.16.2 # --index-url https://download.pytorch.org/whl/cu118
torchaudio==2.1.2   # --index-url https://download.pytorch.org/whl/cu118
xgboost==2.0.3
pandas_market_calendars==4.3.3
yfinance==0.2.34
pandas_ta==0.3.14b0  # required python 3.9 as of 12-04-2024
torchinfo==1.8.0
tensorboard==2.15.1
packaging==23.2
scikit-learn==1.3.2
matplotlib==3.7.3
fastapi==0.109.0
uvicorn[standard]==0.25.0
pymongo[gssapi,srv,tls]==4.6.1